# a7_openapi.yaml
openapi: 3.0.0
info:
 version: '1.0'
 title: 'LBAW WeMeet Web API'
 description: 'Web Resources Specification (A7) for WeMeet'
servers:
- url: https://lbaw22102.lbaw.fe.up.pt/
  description: Production server
...
paths:
 /login:
   get:
     operationId: R101
     summary: 'R101: Login Form'
     description: 'Provide login form. Access: PUB'
     tags:
       - 'M01: Authentication and Individual Profile'
     responses:
       '200':
         description: 'Ok. Show [UI06]'
   post:
     operationId: R102
     summary: 'R102: Login Action'
     description: 'Processes the login form submission. Access: PUB'
     tags:
       - 'M01: Authentication and Individual Profile'

     requestBody:
       required: true
       content:
         application/x-www-form-urlencoded:
           schema:
             type: object
             properties:
               password:    
                 type: string
               email:          
                 type: string
             required:
                  - password

     responses:
       '302':
         description: 'Redirect after processing the login credentials.'
         headers:
           Location:
             schema:
               type: string
             examples:
               302Success:
                 description: 'Successful authentication. Redirect to events page.'
                 value: '/events'
               302Error:
                 description: 'Failed authentication. Redirect to login form.'
                 value: '/login'
...
 /event:
    get:
      operationId: R209
      summary: "R209: Search Events"
      description: "Searches for events and returns the results as JSON. Access: PUB"
      tags:
        - 'M02: Events'

      parameters:
        - in: query
          name: query
          description: String to use for full-text search
          schema:
            type: string
          required: false
        - in: query
          name: closed
          description: Boolean with the closed flag value
          schema:
            type: boolean
          required: false

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Bad Request
...
components:
  schemas:
    Event:
        type: object
        properties:
          id:
            type: integer
          title:
            type: string
          description:
            type: string
          visibility:
            type: boolean
          picture:
            type: string
            format: byte
          publish_date:
            type: string 
          start_date:
            type: string
          final_sate:
            type: string
          is_canceled:
            type: boolean